requirements.txt
requests
pandas
pandas-ta
python-binance
binance-connector
websocket-client
